requirements.txt
amqp==5.2.0
asgiref==3.7.2
azure-core==1.29.5
azure-digitaltwins-core==1.2.0
azure-identity==1.15.0
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.5
Django==4.2.7
djangorestframework==3.14.0
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
isodate==0.6.1
kombu==5.3.4
msal==1.25.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
packaging==23.2
portalocker==2.8.2
prompt-toolkit==3.0.41
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
